dav: Add webdav handler to /dav/
This is WORK IN PROGRESS, DO NOT MERGE, MVP, experiment.

Uses pkg/fs which is very quirky, as that implements bazil.org/fuse,
golang.org/x/net/webdav requires http.File, so that now is served
through bazil.org/fuse.

The nicer solution would be to carve out an fs.FS implementation from
pkg/fs, and implement fuse on top of it.