diskpacked: limit max open files with SIEVE cache
It has opened (and kept open) all pack-%05d.blobs file.

With this change, a SIEVE cache is used with 80% of MaxFDs (or of 1024).
Why SIEVE?
For the random access pattern we mostly face with,
with an 1/8 size cache (1024 cache size, 8192 different items),
  - LRU and random has 1% hit rate,
  - SIEVE has 8%
(out of the optimal 1/8=12.5%).

The sieve code has been copied from
https://pkg.go.dev/github.com/opencoff/go-sieve
(BSD-2) and customized:
  - removed mutexes (here fdCache has a global mutex),
  - removed not really needed functions (Dump, String and Purge),
  - added RemoveOldest (mimic LRU) - for this
    let evict return the evicted node.
  - added a callback function to be called on each
    removed (evicted) item - here we use it to close the file.